Move file *moved PositionCalculator.java to utils
